requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.86.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.31
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.19.0
